+++ frontend/src/components/home.component.js
@@ -1,6 +1,7 @@
 /* eslint-disable class-methods-use-this */
 /* eslint-disable no-unused-vars */
 import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
 
 
 
@@ -13,7 +14,7 @@ export default class Home extends Component {
                 <div className="col-sm-12 text-center py-5">
                     <h1>Welcome to GoBusiness</h1>
                     <h4>A Platform for SMEs Investment</h4>
-                      <a href="/signup"><button className="btn btn-primary px-5 py-2 my-3 start shadow">Get Started</button></a>
+                    <Link to="/signup"><button className="btn btn-primary px-5 py-2 my-3 start shadow">Get Started</button></Link>
                 </div>
           
                 </div>
